Refactor `viewNotes.php` API endpoint
