+++ app/Models/BuildNote.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 /**
  * @mixin Builder<BuildNote>
@@ -19,6 +20,8 @@ class BuildNote extends Model
 
     /**
      * Get the note for this buildnote
+     *
+     * @return BelongsTo<Note, self>
      */
     public function note()
     {
